+++ tests/frameservers/tui_test/tui_test.c
@@ -212,7 +212,7 @@ static void on_mouse(struct tui_context* c,
 }
 
 static void on_key(struct tui_context* c, uint32_t xkeysym,
-	uint8_t scancode, uint8_t mods, uint16_t subid, void* t)
+	uint8_t scancode, uint16_t mods, uint16_t subid, void* t)
 {
 	trace("unknown_key(%"PRIu32",%"PRIu8",%"PRIu16")", xkeysym, scancode, subid);
 }
